+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [1.7.1] - 2025-04-13
+
+### Added
+- The `DASSLCIPHERS` environment variable and `nginx ssl ciphers`
+  Configuration directive for controlling the `ssl_ciphers` NGINX
+  configuration option.
+### Fixed
+- Problem when `valid formats` with `code` was used with `docx
+  template file`.
+
 ## [1.7.0] - 2025-04-09
 
 ### Added